Rename pivoted table to cluster_members_per_ring_mean
Previously, the result of applying the pandas .pivot_table() function
was stored in a table with the same name (cluster_members_per_ring),
which could cause confusion.

The new table is now named cluster_members_per_ring_mean, since in this
case .pivot_table() computes the mean for each column.